+++ app/_layout.tsx
@@ -5,7 +5,9 @@ import { Stack } from 'expo-router';
 import * as SplashScreen from 'expo-splash-screen';
 import { useEffect } from 'react';
 
-import { useColorScheme } from '@/components/useColorScheme';
+import { useColorScheme } from '@/hooks/useColorScheme';
+import {Provider} from "react-redux";
+import {store} from "@/store";
 export { ErrorBoundary } from 'expo-router';
 
 export const unstable_settings = {
@@ -45,9 +47,11 @@ function RootLayoutNav() {
 
   return (
     <ThemeProvider value={colorScheme === 'dark' ? DarkTheme : DefaultTheme}>
+      <Provider store={store}>
       <Stack>
         <Stack.Screen name="(tabs)" options={{ headerShown: false }} />
       </Stack>
+      </Provider>
     </ThemeProvider>
   );
 }
