+++ components/EditScreenInfo.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { StyleSheet } from 'react-native';
 
 import { ExternalLink } from './ExternalLink';
-import { MonoText } from './StyledText';
+import { MonoText } from './StylesText/StyledText';
 import { Text, View } from './Themed';
 
 import Colors from '@/constants/Colors';
