+++ components/StylesText/__tests__/StyledText.test.js
