+++ package.json
@@ -15,6 +15,7 @@
   "dependencies": {
     "@expo/vector-icons": "^14.0.0",
     "@react-navigation/native": "^6.0.2",
+    "@reduxjs/toolkit": "^2.2.1",
     "expo": "~50.0.7",
     "expo-font": "~11.10.3",
     "expo-linking": "~6.2.2",
